% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2oaiglm.R
\name{coef.h2oaiglmnet}
\alias{coef.h2oaiglmnet}
\title{h2oaiglmnet coefficients.}
\usage{
\method{coef}{h2oaiglmnet}(object, s = NULL, ...)
}
\arguments{
\item{object}{Value returned by h2oaiglmnet(..).}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Returns coeficients of h2oaiglmnet fit.
}

