% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2oaiglm.R
\name{kNegEntr}
\alias{kNegEntr}
\title{Negative Entropy Function}
\usage{
kNegEntr(m = 1)
}
\arguments{
\item{m}{Number of times that the function should be repeated.}
}
\description{
\deqn{h(x) = x log(x)}
}

