% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2oaiglm.R
\name{plot.h2oaiglmnet}
\alias{plot.h2oaiglmnet}
\title{h2oaiglmnet coefficient plot.}
\usage{
\method{plot}{h2oaiglmnet}(x, xvar = c("norm", "lambda", "dev"), ...)
}
\arguments{
\item{x}{h2oaiglmnet object.}

\item{xvar}{x-variable to plot coefficient against.}

\item{...}{Extra parameters for matplot.}
}
\description{
Plots coefficients
}

