% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2oaiglm.R
\name{print.h2oaiglmnet}
\alias{print.h2oaiglmnet}
\title{Prints h2oaiglmnet summary}
\usage{
\method{print}{h2oaiglmnet}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{h2oaiglmnet object}

\item{digits}{Number of digits to display}

\item{...}{Not used.}
}
\description{
Prings h2oaiglmnet summary
}

